requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
requests==2.31.0
duckduckgo-search==4.1.1
aiohttp==3.9.1
pytest==7.4.3
pytest-asyncio==0.21.1
flake8==6.1.0
flask==3.0.2
flask-login==0.6.3
flask-sqlalchemy==3.1.1
werkzeug==3.0.1 